ironing out the schema implementation (colour picker)
