+++ components/blocks/sharedTinaFields/colourOptions/blockBackgroundOptions.tsx
@@ -1,0 +1,28 @@
+import { ColorPickerOptions } from "../colourSelector";
+
+export const backgroundOptions: ColorPickerOptions[] = [
+  {
+    name: "Soft Left Gradient",
+    classes: "bg-gradient-to-l from-gray-900 to-gray-1000 text-white",
+  },
+  {
+    name: "Soft Right Gradient",
+    classes: "bg-gradient-to-r from-gray-900 to-gray-1000 text-white",
+  },
+  {
+    name: "Sheer Top Gradient",
+    classes: "bg-gradient-to-t from-gray-900 to-black text-white",
+  },
+  {
+    name: "Sheer Bottom Gradient",
+    classes: "bg-gradient-to-b from-gray-900 to-black text-white",
+  },
+  {
+    name: "Transparent",
+    classes: "bg-transparent text-gray-950",
+  },
+  {
+    name: "Gray",
+    classes: "bg-gray-700 text-white",
+  },
+];
