+++ components/blocks/sharedTinaFields/colourOptions/buttonOptions.tsx
@@ -1,0 +1,12 @@
+import { ColorPickerOptions } from "../colourSelector";
+
+export const buttonOptions: ColorPickerOptions[] = [
+  {
+    name: "Red",
+    classes: "bg-red-700 text-white",
+  },
+  {
+    name: "Transparent",
+    classes: "bg-transparent text-gray-950",
+  },
+];
